THO-167 Bug/stack windows (#22)
* directionalLight when is nullptr fixed

* bug fixed and menu bar polished

* all build warnings fixed from first release

+++ SobrassadaEngine/ResourceManagement/Resources/ResourceMesh.cpp
@@ -34,8 +34,8 @@ void ResourceMesh::LoadData(
 {
     this->mode = mode;
     this->material = material;
-    this->vertexCount       = vertices.size();
-    this->indexCount        = indices.size();
+    this->vertexCount       = static_cast<unsigned int>(vertices.size());
+    this->indexCount        = static_cast<unsigned int>(indices.size());
     unsigned int bufferSize = sizeof(Vertex);
 
     glGenVertexArrays(1, &vao);
